fix: signed post-repair merge early-stop compare + shelf-tier fixture tripwire
AtlasCharting.cpp: the post-repair merge early-stop compared before-numCharts
against before/100 as unsigned — a round that increases numCharts underflows
the subtraction. Cast to long so the comparison is signed, preserving the
normal (decreasing) semantics.

AtlasTest.cpp: TwoTierManyTinyChartsDisjointAndDense asserted the packed
result was disjoint and dense but never proved its 2000 tiny charts actually
landed in the shelf tier (AtlasPacking.cpp's tierThreshold = pageW/32).
Density-normalization drift could silently push them all into the skyline
tier and the test would keep passing without covering the shelf path. Add a
fixture-validity assertion that >90% of the tiny charts fall below the
padded pageW/32 threshold post-pack; verified the assertion actually fires
by temporarily starving the fixture's chart-count disparity and confirming
the new EXPECT_GT fails before reverting.

+++ src/AtlasCharting.cpp
@@ -1308,7 +1308,7 @@ unsigned SegmentCharts(Mesh& mesh, const ParametrizeParams& params,
 #ifdef HM_ATLAS_DEBUG
 			std::cerr << "[re-merge] round " << round << ": " << before << " -> " << numCharts << " charts\n";
 #endif
-			if (before - numCharts < before / 100)
+			if (static_cast<long>(before) - static_cast<long>(numCharts) < static_cast<long>(before / 100))
 				break; // <1% net change — not worth another round
 		}
 	}
